+++ test/base/artifact/test_geotiff.py
@@ -1,6 +1,5 @@
 import logging
 import os
-from pathlib import Path
 
 import numpy as np
 import pytest
@@ -21,16 +20,7 @@ from climatoology.base.artifact import (
 )
 
 
-def test_create_concise_geotiff_artifact_2d(default_computation_resources, general_uuid):
-    expected_artifact = _Artifact(
-        name='Test Raster',
-        modality=ArtifactModality.MAP_LAYER_GEOTIFF,
-        filename='test_file.tiff',
-        summary='Raster caption',
-        attachments=Attachments(legend=Legend(legend_data={'1': Color('#0f0')})),
-        correlation_uuid=general_uuid,
-    )
-
+def test_create_concise_geotiff_artifact_2d(default_computation_resources, default_artifact, default_artifact_metadata):
     method_input = RasterInfo(
         data=np.ones(shape=(4, 5), dtype=np.uint8),
         crs=CRS({'init': 'epsg:4326'}),
@@ -42,41 +32,27 @@ def test_create_concise_geotiff_artifact_2d(default_computation_resources, gener
             d=0.0,
             e=0.1,
         ),
-        colormap=colormap_type({1: (0, 255, 0)}),
     )
+
+    default_artifact_copy = default_artifact.model_copy(deep=True)
+    default_artifact_copy.modality = ArtifactModality.MAP_LAYER_GEOTIFF
+    default_artifact_copy.filename = f'{default_artifact_metadata.filename}.tiff'
 
     generated_artifact = create_geotiff_artifact(
         raster_info=method_input,
-        layer_name='Test Raster',
-        caption='Raster caption',
+        metadata=default_artifact_metadata,
         resources=default_computation_resources,
-        filename='test_file',
     )
+    with rasterio.open(default_computation_resources.computation_dir / generated_artifact.filename) as generated_file:
+        generated_content = generated_file.read()
 
-    assert generated_artifact == expected_artifact
-
-    with rasterio.open(
-        default_computation_resources.computation_dir / generated_artifact.filename
-    ) as generated_content:
-        assert_array_equal(generated_content.read(), [method_input.data])
+    assert generated_artifact == default_artifact_copy
+    assert_array_equal(generated_content, [method_input.data])
 
 
 def test_create_extensive_geotiff_artifact_2d(
-    default_computation_resources, general_uuid, default_association_tags, default_sources
+    default_computation_resources, extensive_artifact, extensive_artifact_metadata
 ):
-    expected_artifact = _Artifact(
-        name='Test Raster',
-        modality=ArtifactModality.MAP_LAYER_GEOTIFF,
-        filename='test_file.tiff',
-        summary='Raster caption',
-        description='Raster description',
-        sources=default_sources,
-        tags=default_association_tags,
-        primary=False,
-        attachments=Attachments(legend=Legend(legend_data={'1': Color('#0f0')}, title='Custom Legend Title')),
-        correlation_uuid=general_uuid,
-    )
-
     method_input = RasterInfo(
         data=np.ones(shape=(4, 5), dtype=np.uint8),
         crs=CRS({'init': 'epsg:4326'}),
@@ -90,23 +66,20 @@ def test_create_extensive_geotiff_artifact_2d(
         ),
         colormap=colormap_type({1: (0, 255, 0)}),
     )
+    legend = Legend(legend_data={'1': Color('#0f0')}, title='Custom Legend Title')
     method_input_copy = method_input.model_copy(deep=True)
 
-    legend = Legend(legend_data={'1': Color('#0f0')}, title='Custom Legend Title')
+    extensive_artifact_copy = extensive_artifact.model_copy(deep=True)
+    extensive_artifact_copy.modality = ArtifactModality.MAP_LAYER_GEOTIFF
+    extensive_artifact_copy.filename = f'{extensive_artifact_metadata.filename}.tiff'
+    extensive_artifact_copy.attachments = Attachments(legend=legend)
+
     generated_artifact = create_geotiff_artifact(
         raster_info=method_input,
-        layer_name='Test Raster',
-        caption='Raster caption',
+        metadata=extensive_artifact_metadata,
         resources=default_computation_resources,
-        primary=False,
         legend=legend,
-        description='Raster description',
-        sources=Path(__file__).parent.parent.parent / 'resources/minimal.bib',
-        tags=default_association_tags,
-        filename='test_file',
     )
-
-    assert generated_artifact == expected_artifact
 
     # Method input should not be mutated during artifact creation
     assert (method_input.data == method_input_copy.data).all()
@@ -114,18 +87,15 @@ def test_create_extensive_geotiff_artifact_2d(
     assert method_input.transformation == method_input_copy.transformation
     assert method_input.colormap == method_input_copy.colormap
 
-    with rasterio.open(
-        default_computation_resources.computation_dir / generated_artifact.filename
-    ) as generated_content:
-        assert_array_equal(generated_content.read(), [method_input.data])
+    assert generated_artifact == extensive_artifact_copy
 
 
-def test_create_geotiff_artifact_2d_rgba(default_computation_resources, general_uuid):
+def test_create_geotiff_artifact_2d_rgba(default_computation_resources, general_uuid, default_artifact_metadata):
     expected_artifact = _Artifact(
-        name='Test Raster',
+        name='test_name',
         modality=ArtifactModality.MAP_LAYER_GEOTIFF,
-        filename='test_file.tiff',
-        summary='Raster caption',
+        filename='test_artifact_file.tiff',
+        summary='Test summary',
         attachments=Attachments(legend=Legend(legend_data={'1': Color('#00ff00fe')})),
         correlation_uuid=general_uuid,
     )
@@ -146,10 +116,8 @@ def test_create_geotiff_artifact_2d_rgba(default_computation_resources, general_
 
     generated_artifact = create_geotiff_artifact(
         method_input,
-        layer_name='Test Raster',
-        caption='Raster caption',
+        metadata=default_artifact_metadata,
         resources=default_computation_resources,
-        filename='test_file',
     )
 
     assert generated_artifact == expected_artifact
@@ -160,7 +128,7 @@ def test_create_geotiff_artifact_2d_rgba(default_computation_resources, general_
         assert_array_equal(generated_content.read(), [method_input.data])
 
 
-def test_create_geotiff_artifact_masked_array(default_computation_resources):
+def test_create_geotiff_artifact_masked_array(default_computation_resources, default_artifact_metadata):
     method_input = RasterInfo(
         data=np.ma.masked_array(np.ones(shape=(2, 2), dtype=np.uint8), mask=[[0, 0], [0, 1]]),
         crs=CRS({'init': 'epsg:4326'}),
@@ -177,10 +145,8 @@ def test_create_geotiff_artifact_masked_array(default_computation_resources):
 
     generated_artifact = create_geotiff_artifact(
         method_input,
-        layer_name='Test Raster',
-        caption='Raster caption',
+        metadata=default_artifact_metadata,
         resources=default_computation_resources,
-        filename='test_file',
     )
 
     with rasterio.open(
@@ -189,12 +155,12 @@ def test_create_geotiff_artifact_masked_array(default_computation_resources):
         assert_array_equal(generated_content.read(1, masked=True), method_input.data)
 
 
-def test_create_geotiff_with_legend_data(default_computation_resources, general_uuid):
+def test_create_geotiff_with_legend_data(default_computation_resources, general_uuid, default_artifact_metadata):
     expected_artifact = _Artifact(
-        name='Test Raster',
+        name='test_name',
         modality=ArtifactModality.MAP_LAYER_GEOTIFF,
-        filename='test_file.tiff',
-        summary='Raster caption',
+        filename='test_artifact_file.tiff',
+        summary='Test summary',
         attachments=Attachments(legend=Legend(legend_data={'A': Color('#f00')})),
         correlation_uuid=general_uuid,
     )
@@ -215,10 +181,8 @@ def test_create_geotiff_with_legend_data(default_computation_resources, general_
 
     generated_artifact = create_geotiff_artifact(
         method_input,
-        layer_name='Test Raster',
-        caption='Raster caption',
+        metadata=default_artifact_metadata,
         resources=default_computation_resources,
-        filename='test_file',
         legend=Legend(legend_data={'A': Color('red')}),
     )
 
@@ -230,12 +194,12 @@ def test_create_geotiff_with_legend_data(default_computation_resources, general_
         assert_array_equal(generated_content.read(), [method_input.data])
 
 
-def test_create_geotiff_without_legend(default_computation_resources, general_uuid):
+def test_create_geotiff_without_legend(default_computation_resources, general_uuid, default_artifact_metadata):
     expected_artifact = _Artifact(
-        name='Test Raster',
+        name='test_name',
         modality=ArtifactModality.MAP_LAYER_GEOTIFF,
-        filename='test_file.tiff',
-        summary='Raster caption',
+        filename='test_artifact_file.tiff',
+        summary='Test summary',
         correlation_uuid=general_uuid,
     )
 
@@ -254,10 +218,8 @@ def test_create_geotiff_without_legend(default_computation_resources, general_uu
 
     generated_artifact = create_geotiff_artifact(
         method_input,
-        layer_name='Test Raster',
-        caption='Raster caption',
+        metadata=default_artifact_metadata,
         resources=default_computation_resources,
-        filename='test_file',
     )
 
     assert generated_artifact == expected_artifact
@@ -268,12 +230,12 @@ def test_create_geotiff_without_legend(default_computation_resources, general_uu
         assert_array_equal(generated_content.read(), [method_input.data])
 
 
-def test_create_geotiff_artifact_3d(default_computation_resources, general_uuid):
+def test_create_geotiff_artifact_3d(default_computation_resources, general_uuid, default_artifact_metadata):
     expected_artifact = _Artifact(
-        name='Test Raster',
+        name='test_name',
         modality=ArtifactModality.MAP_LAYER_GEOTIFF,
-        filename='test_file.tiff',
-        summary='Raster caption',
+        filename='test_artifact_file.tiff',
+        summary='Test summary',
         attachments=Attachments(legend=Legend(legend_data={'1': Color('#0f0')})),
         correlation_uuid=general_uuid,
     )
@@ -294,10 +256,8 @@ def test_create_geotiff_artifact_3d(default_computation_resources, general_uuid)
 
     generated_artifact = create_geotiff_artifact(
         method_input,
-        layer_name='Test Raster',
-        caption='Raster caption',
+        metadata=default_artifact_metadata,
         resources=default_computation_resources,
-        filename='test_file',
     )
 
     assert generated_artifact == expected_artifact
@@ -308,7 +268,7 @@ def test_create_geotiff_artifact_3d(default_computation_resources, general_uuid)
         assert_array_equal(generated_content.read(), method_input.data)
 
 
-def test_geotiff_creation_option_warnings(default_computation_resources, caplog):
+def test_geotiff_creation_option_warnings(default_computation_resources, caplog, default_artifact_metadata):
     method_input = RasterInfo(
         data=np.ones(shape=(4, 5), dtype=np.uint8),
         crs=CRS({'init': 'epsg:4326'}),
@@ -326,16 +286,14 @@ def test_geotiff_creation_option_warnings(default_computation_resources, caplog)
     with caplog.at_level(logging.INFO):
         create_geotiff_artifact(
             method_input,
-            layer_name='Test Raster',
-            caption='Raster caption',
+            metadata=default_artifact_metadata,
             resources=default_computation_resources,
-            filename='test_file',
         )
 
     assert caplog.messages == []
 
 
-def test_small_geotiff_created_without_overviews(default_computation_resources):
+def test_small_geotiff_created_without_overviews(default_computation_resources, default_artifact_metadata):
     method_input = RasterInfo(
         data=np.ones(shape=(10, 10), dtype=np.uint8),
         crs=CRS({'init': 'epsg:4326'}),
@@ -351,10 +309,8 @@ def test_small_geotiff_created_without_overviews(default_computation_resources):
 
     generated_artifact = create_geotiff_artifact(
         method_input,
-        layer_name='Test Raster',
-        caption='Raster caption',
+        metadata=default_artifact_metadata,
         resources=default_computation_resources,
-        filename='test_file',
     )
 
     with rasterio.open(
@@ -363,7 +319,7 @@ def test_small_geotiff_created_without_overviews(default_computation_resources):
         assert generated_content.overviews(1) == []
 
 
-def test_geotiff_created_with_overviews(default_computation_resources):
+def test_geotiff_created_with_overviews(default_computation_resources, default_artifact_metadata):
     method_input = RasterInfo(
         data=np.ones(shape=(1000, 1000), dtype=np.uint8),
         crs=CRS({'init': 'epsg:4326'}),
@@ -380,10 +336,8 @@ def test_geotiff_created_with_overviews(default_computation_resources):
 
     generated_artifact = create_geotiff_artifact(
         method_input,
-        layer_name='Test Raster',
-        caption='Raster caption',
+        metadata=default_artifact_metadata,
         resources=default_computation_resources,
-        filename='test_file',
     )
 
     with rasterio.open(
@@ -392,7 +346,7 @@ def test_geotiff_created_with_overviews(default_computation_resources):
         assert generated_content.overviews(1) == [2, 4]
 
 
-def test_geotiff_tiled(default_computation_resources):
+def test_geotiff_tiled(default_computation_resources, default_artifact_metadata):
     method_input = RasterInfo(
         data=np.ones(shape=(1000, 1000), dtype=np.uint8),
         crs=CRS({'init': 'epsg:4326'}),
@@ -409,10 +363,8 @@ def test_geotiff_tiled(default_computation_resources):
 
     generated_artifact = create_geotiff_artifact(
         method_input,
-        layer_name='Test Raster',
-        caption='Raster caption',
+        metadata=default_artifact_metadata,
         resources=default_computation_resources,
-        filename='test_file',
     )
 
     with rasterio.open(
@@ -423,7 +375,7 @@ def test_geotiff_tiled(default_computation_resources):
         assert generated_content.profile['blockysize'] == 512
 
 
-def test_disallow_colormap_for_incompatible_dtype(default_computation_resources):
+def test_disallow_colormap_for_incompatible_dtype(default_computation_resources, default_artifact_metadata):
     method_input = RasterInfo(
         data=np.ones(shape=(4, 5), dtype=float),
         crs=CRS({'init': 'epsg:4326'}),
@@ -441,8 +393,7 @@ def test_disallow_colormap_for_incompatible_dtype(default_computation_resources)
     with pytest.raises(AssertionError, match='Colormaps are not allowed for dtype float.'):
         create_geotiff_artifact(
             method_input,
-            layer_name='Test Raster',
-            caption='Raster caption',
+            metadata=default_artifact_metadata,
             resources=default_computation_resources,
         )