+++ test/base/artifact/test_markdown.py
@@ -1,67 +1,43 @@
-from pathlib import Path
-
-from climatoology.base.artifact import ArtifactModality, _Artifact, create_markdown_artifact
+from climatoology.base.artifact import create_markdown_artifact
 
 
-def test_create_concise_markdown_artifact(default_computation_resources, general_uuid):
-    expected_artifact = _Artifact(
-        name='-',
-        modality=ArtifactModality.MARKDOWN,
-        filename='test_file.md',
-        summary='-',
-        correlation_uuid=general_uuid,
-    )
-    expected_content = """# Header
+def test_create_concise_markdown_artifact(default_computation_resources, default_artifact, default_artifact_metadata):
+    method_input = """# Header
 
     Content
     """
 
     generated_artifact = create_markdown_artifact(
-        text=expected_content,
+        text=method_input,
+        metadata=default_artifact_metadata,
         resources=default_computation_resources,
-        name='-',
-        tl_dr='-',
-        filename='test_file',
     )
     with open(default_computation_resources.computation_dir / generated_artifact.filename, 'r') as test_file:
         generated_content = test_file.read()
 
-    assert generated_artifact == expected_artifact
-    assert generated_content == expected_content
+    assert generated_artifact == default_artifact
+    assert generated_content == method_input
 
 
 def test_create_extensive_markdown_artifact(
-    default_computation_resources, general_uuid, default_association_tags, default_sources
+    default_computation_resources, general_uuid, extensive_artifact, default_sources, extensive_artifact_metadata
 ):
-    expected_artifact = _Artifact(
-        name='-',
-        modality=ArtifactModality.MARKDOWN,
-        filename='test_file.md',
-        summary='-',
-        sources=default_sources,
-        primary=False,
-        tags=default_association_tags,
-        correlation_uuid=general_uuid,
-    )
-    expected_content = """# Header
+    method_input = """# Header
 
     Content
     """
-    expected_content_copy = expected_content
+    method_input_copy = method_input
+
+    extensive_artifact_copy = extensive_artifact.model_copy(deep=True)
+    extensive_artifact_copy.description = None
+    extensive_artifact_metadata_copy = extensive_artifact_metadata.model_copy(deep=True)
+    extensive_artifact_metadata_copy.description = None
 
     generated_artifact = create_markdown_artifact(
-        text=expected_content,
+        text=method_input,
+        metadata=extensive_artifact_metadata_copy,
         resources=default_computation_resources,
-        name='-',
-        tl_dr='-',
-        sources=Path(__file__).parent.parent.parent / 'resources/minimal.bib',
-        primary=False,
-        tags=default_association_tags,
-        filename='test_file',
     )
-    with open(default_computation_resources.computation_dir / generated_artifact.filename, 'r') as test_file:
-        generated_content = test_file.read()
 
-    assert expected_content_copy == expected_content
-    assert generated_artifact == expected_artifact
-    assert generated_content == expected_content
+    assert method_input_copy == method_input, 'Method input should not be mutated during artifact creation'
+    assert generated_artifact == extensive_artifact_copy